Add files via upload
Decided to use datamuse for the api.

+++ script.js
@@ -1,29 +1,24 @@
 
 
 
-
-options = {
-	method: 'GET',
-	headers: {
-		"Accept": "application/json","app_id": "0d5e9a2e", "app_key": "0873c0ac9a8c2ec4df73424a9df89768"
-	}
-};
-
 const listsDiv = document.getElementById('listsDiv');
 const historyDiv = document.getElementById('historyDiv');
-const synListUl = document.getElementById('synListUl');
-const antListUl = document.getElementById('antListUl');
+const synStat = document.getElementById('synStat');
+const synListDiv = document.getElementById('synListDiv')
+let synListUl = document.getElementById('synListUl');
+const antStat = document.getElementById('antStat');
+const antListDiv = document.getElementById('antListDiv');
+let antListUl = document.getElementById('antListUl');
 const historyListUl = document.getElementById('historyListUl');
 const submit = document.getElementById('submit');
 const clear = document.getElementById('clear');
 const searchInput = document.getElementById('searchInput');
 const searchWord = document.getElementById('searchWord');
 const images = document.querySelectorAll('img');
-
 let word;
 let data;
-let synonymsList;
-let antonymsList;
+let synonymsList = [];
+let antonymsList = [];
 let historyList = [];
 
 
@@ -35,8 +30,11 @@ if (localStorage.length > 0) {
 
 // clears local storage
 clear.onclick = function() {
-    localStorage.clear();
-    location.reload();
+    if(confirm('CLEAR? This will clear everything!')) {
+        localStorage.clear();
+        location.reload();
+    }
+    
 }
 
 // triggers submit 'click' if 'enter' is pressed
@@ -57,31 +55,43 @@ submit.onclick = function() {
 
 // fetches the searched word from a thesaurus database api
 function fetchData(word) {
-    fetch(`https://od-api.oxforddictionaries.com/api/v2/thesaurus/en/${word}?strictMatch=false`, options)
-	.then(response => response.json())
+    Promise.all([
+        fetch(`https://api.datamuse.com/words?rel_syn=${word}`),
+        fetch(`https://api.datamuse.com/words?rel_ant=${word}`)
+    ])
+	.then(function (responses) {
+        // Get a JSON object from each of the responses
+        return Promise.all(responses.map(function (response) {
+            return response.json();
+        }))
+    })
 	.then(function(response){
         searchWord.innerHTML = word;
         data = response;
+        
+        synonymsList = data[0];
+        antonymsList = data[1];
 
-        synonymsList = data.results[0].lexicalEntries[0].entries[0].senses[0].synonyms;
- 
-        synonymsList.forEach(createSynonymsList)
+        createList(synListUl, synonymsList, 0, synonymsList.length);
+        createList(antListUl, antonymsList, 0, antonymsList.length)
 
-        if(data.results[0].lexicalEntries[0].entries[0].senses[0].antonyms.length > 0) {
-            antonymsList = data.results[0].lexicalEntries[0].entries[0].senses[0].antonyms;
-            antonymsList.forEach(createAntonymsList)
-        }
-        // console.log(synonymsList)
+        synStat.innerHTML = `${synonymsList.length}`;
+        antStat.innerHTML = `${antonymsList.length}`;
+
         historyList.forEach(createHistoryList);
+
+        hide('synListUl');
+        
         clickWord();
         save();
     } )
 	.catch(err => console.error(err));
 };
 
-// creates the click listener for list items and their copy icons
+// creates the click listener for list items and their copy icons and expanding lists
 function clickWord() {
     listsDiv.onclick = function(event) {
+        // event.preventDefault();
         if(event.target.nodeName == 'LI') {
             word = event.target.innerHTML;
             historyListArray();
@@ -91,7 +101,23 @@ function clickWord() {
         if(event.target.nodeName == 'IMG') {
             navigator.clipboard.writeText(event.target.previousSibling.innerHTML);
         }
+        // console.log(event.target.classList.contains('more'))
+        // expands lists if arrow is clicked
+        if(event.target.classList.contains('more')){
+            // console.log(event.target.parentElement.id)
+            if(event.target.parentElement.id == 'synListDiv'){
+                show('synListUl');
+                toggleMore(synListUl, synListDiv);
+            }
+        }
+        if(event.target.classList.contains('less')){
+            if(event.target.parentElement.id == 'synListDiv'){
+                hide('synListUl');
+                toggleMore(synListUl, synListDiv);
+            }
+        }
     }
+    // adds copy function to history items
     historyDiv.onclick = function(event) {
         if(event.target.nodeName == 'LI') {
             word = event.target.innerHTML;
@@ -113,7 +139,7 @@ function historyListArray() {
             dupe = true;
         }
     });
-    // console.log(dupe);
+// list is capped at 5 - adds to the end and removes the first
     if(historyList.length < 5 && dupe == false) {
         historyList.push(word);
     }else if(dupe == false){
@@ -128,23 +154,72 @@ function createHistoryList(item) {
     `
 };
 
-function createSynonymsList(item) {
-    console.log(item.text);
-    synListUl.innerHTML += `
-        <div class="list-item"><li>${item.text}</li><img src="images/copy-icon.png" alt="[ ]" title="copy ${item.text}"></div>
-    `
+// creates list for Synonyms and Antonyms depending on which UL is passed
+function createList(ul, array,  start, end) {
+    for(let i = start; i < end; i++){
+        ul.innerHTML += `
+            <div class="list-item"><li>${array[i].word}</li><img src="images/copy-icon.png" alt="[ ]" title="copy ${array[i].word}"></div>
+        `
+    }
+    listItems = document.querySelectorAll(`#${ul.id} .list-item`);
+
+    if(listItems.length > 20){
+        synListDiv.innerHTML += `
+            <p id="${ul.id}More" class="more">more...</p>
+        `
+    }
 };
 
-function createAntonymsList(item) {
-    antListUl.innerHTML += `
-        <div class="list-item"><li>${item.text}</li><img src="images/copy-icon.png" alt="[ ]" title="copy ${item.text}"></div>
-    `
-};
+// toggles the more... less... link if the list has more than 20 words
+function toggleMore(ul, listDiv) {
+    if(document.getElementById(`${ul.id}More`)){
+        // console.log(true)
+        let listUlMore = document.getElementById(`${ul.id}More`);
+        listUlMore.remove();
+        listDiv.innerHTML += `
+            <p id="${ul.id}Less" class="less">less...</p>
+        `
+    }else{
+        let listUlLess = document.getElementById(`${ul.id}Less`);
+        listUlLess.remove();
+        listDiv.innerHTML += `
+            <p id="${ul.id}More" class="more">more...</p>
+        `
+    }
+}
 
+// hides the rest of the list if the list goes over 20
+function hide(ul) {
+    listItems = document.querySelectorAll(`#${ul} .list-item`);
+    if(ul)
+    for(let i=20; i < listItems.length; i++) {
+        listItems[i].classList.add('hide');
+    }
+}
+
+// shows the rest of the list over 20 to the end of list
+function show(ul) {
+    listItems = document.querySelectorAll(`#${ul} .list-item`);
+    for(let i=20; i < listItems.length; i++) {
+        listItems[i].classList.remove('hide');
+    }
+}
+
+// clears all lists and more links
 function listsClear() {
+    synListUl = document.getElementById('synListUl');
+    antListUl = document.getElementById('antListUl');
     synListUl.innerHTML = '';
     antListUl.innerHTML = '';
     historyListUl.innerHTML = '';
+    let more = document.querySelectorAll('.more');
+    more.forEach(function(item){
+        item.remove()
+    })
+    let less = document.querySelectorAll('.less');
+    less.forEach(function(item){
+        item.remove()
+    })
 }
 
 // saves current word and history to local storage
@@ -158,6 +233,7 @@ function load() {
     word = localStorage.getItem('word');
     historyList = JSON.parse(localStorage.getItem('historyList'));
 };
+
